chore(client): rename vm service in aurae client
